Finished code for calculations, but still there is a bug with wrong calculated values. For now, only speedup is valid.

+++ c_implementation/weights/network2/b1.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (2048,)
-extern double b1[2048];
+extern float b1[2048];
