+++ c_implementation/weights/network2/b2.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (1024,)
-extern double b2[1024];
+extern float b2[1024];
