+++ c_implementation/weights/network2/b3.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (512,)
-extern double b3[512];
+extern float b3[512];
