+++ c_implementation/weights/network2/bout.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (10,)
-extern double bout[10];
+extern float bout[10];
