+++ c_implementation/weights/network2/osmica.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (784,)
-extern double osmica[784];
+extern float osmica[784];
