+++ c_implementation/weights/network2/w2.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (2048, 1024)
-extern double w2[1024][2048];
+extern float w2[1024][2048];
