+++ c_implementation/weights/network2/wout.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (512, 10)
-extern double wout[10][512];
+extern float wout[10][512];
